refactor(Jackpot, Match, Ticket): 주요 기능을 담당하는 클래스를 modules 디렉토리로 이동하여 구조 개선
