+++ src/modules/Ticket.js
@@ -1,18 +1,20 @@
 // 사용자에게 로또 구매 금액을 입력받고 금액만큼 로또 번호를 발행해준다.
 
 import { Console } from '@woowacourse/mission-utils';
-import { INPUT, OUTPUT } from './constants/Constants.js';
-import { ticketCount } from './utils/Calculation.js';
-import { ticketArray } from './utils/TicketNumber.js';
-import Validators from './utils/Validation.js';
+import { INPUT, OUTPUT } from '../constants/Constants.js';
+import { ticketCount } from '../utils/Calculation.js';
+import { ticketArray } from '../utils/TicketNumber.js';
+import Validators from '../utils/Validation.js';
 
 class Ticket {
   #tickets;
   #cost;
+  #lottos;
 
   constructor() {
     this.#tickets = 0;
     this.#cost = 0;
+    this.#lottos = [];
   }
 
   async inputCost() {
@@ -35,8 +37,8 @@ class Ticket {
   }
 
   getTicket() {
-    const Tickets = ticketArray(this.#tickets);
-    return Tickets;
+    this.#lottos = ticketArray(this.#tickets);
+    return this.#lottos;
   }
 
   displayTicket() {
